+++ pkg/webui/console/store/middleware/logics/init.js
@@ -12,28 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { defineMessages } from 'react-intl'
-
 import api from '@console/api'
 
 import * as user from '@console/store/actions/user'
 import * as init from '@console/store/actions/init'
 
 import createRequestLogic from './lib'
-
-const m = defineMessages({
-  errTooFewRights: 'Your account does not possess sufficient rights to use the console',
-  errStateRequested:
-    'Your account still needs to be approved by an administrator. You will receive a confirmation email once your account is approved.',
-  errStateRejected: 'Your account has been rejected by an administrator',
-  errStateSuspended:
-    'Your account has been suspended by an administrator. Please contact support for further information about your account status.',
-  errEmailValidation: 'Your account is restricted until your email address has been validated',
-})
-
-// Define a minimum set of rights, without which it makes no sense to use the
-// console.
-const minimumRights = ['RIGHT_APPLICATION', 'RIGHT_GATEWAY', 'RIGHT_ORGANIZATION']
 
 const consoleAppLogic = createRequestLogic({
   type: init.INITIALIZE,
@@ -65,23 +49,6 @@ const consoleAppLogic = createRequestLogic({
         ])
         userResult.isAdmin = info.is_admin || false
         dispatch(user.getUserMeSuccess(userResult))
-
-        // Check whether the user account has sufficient rights to use the
-        // console.
-        if (!info.is_admin && !rights.some(r => minimumRights.some(mr => r.startsWith(mr)))) {
-          // Provide relevant error messages if possible.
-          if (userResult.state === 'STATE_REQUESTED') {
-            throw m.errStateRequested
-          } else if (userResult.state === 'STATE_REJECTED') {
-            throw m.errStateRejected
-          } else if (userResult.state === 'STATE_SUSPENDED') {
-            throw m.errStateSuspended
-          } else if (!userResult.primary_email_address_validated_at) {
-            throw m.errEmailValidation
-          }
-
-          throw m.errTooFewRights
-        }
       } catch (error) {
         dispatch(user.getUserMeFailure(error))
       }